scraper/requirements.txt
requests
google-genai
supabase
python-telegram-bot
python-dotenv